# Conflicts:
#	build/tmp/compileJava/source-classes-mapping.txt
#	src/main/java/com/spring/mvc/web/michelin/service/michelinService.java
#	src/main/webapp/WEB-INF/views/michelin/list.jsp

+++ src/main/java/com/spring/mvc/web/michelin/domain/Michelin.java
@@ -17,7 +17,7 @@ public class Michelin {
     private String writer;
     private String title;
     private String content;
-    private int star;
+    private double star;
     private Date leg_date;
     private int comment_cnt;
 
@@ -25,7 +25,7 @@ public class Michelin {
         this.board_num = ++sequence;
     }
 
-    public Michelin(String writer, String title, String content, int star) {
+    public Michelin(String writer, String title, String content, double star) {
         this();
         this.writer = writer;
         this.title = title;
