+++ src/main/java/com/spring/mvc/web/michelin/repository/michelinMapper.java
@@ -33,7 +33,7 @@ public interface michelinMapper {
     List<Michelin> starFindAll();
 
     //글 수정
-    void modify(int board_num, String title, String content, String writer, int star);
+    void modify(int board_num, String title, String content, String writer, double star);
 
     //총 게시글 조회기능
     int getCount();
